refactor layout and provider hierarchy
