--- apps/web/src/app/html.tsx
@@ -1,21 +1,0 @@
-import { Inter } from 'next/font/google';
-const inter = Inter({ subsets: ['latin'] });
-
-export default function Html({ children, locale }: { children: React.ReactNode; locale: string }) {
-  return (
-    <html suppressHydrationWarning lang="en">
-      <head>
-        <meta
-          content="Level up your typescript skills with interactive exercises"
-          name="description"
-        />
-        <link href="/favicon.svg" rel="icon" type="image/svg+xml" />
-        <link rel="apple-touch-icon" sizes="180x180" href="/apple-touch-icon.png" />
-        <link rel="icon" type="image/png" sizes="32x32" href="/favicon-32x32.png" />
-        <link rel="icon" type="image/png" sizes="16x16" href="/favicon-16x16.png" />
-        <link rel="manifest" href="/site.webmanifest" />
-      </head>
-      <body className={`${inter.className} flex flex-col`}>{children}</body>
-    </html>
-  );
-}
