+++ apps/web/src/app/layout.tsx
@@ -1,7 +1,17 @@
 import { OG_URL, tagline } from './metadata';
-
+import { getStaticParams } from '~/locales/server';
+import { Inter } from 'next/font/google';
+import { Providers } from './[locale]/providers';
+import { Navigation } from '~/components/navigation';
+import { Toaster } from '@repo/ui/components/toaster';
+import { Analytics } from '@vercel/analytics/react';
 import '../styles/globals.css';
-import Html from './html';
+
+export function generateStaticParams() {
+  return getStaticParams();
+}
+
+const inter = Inter({ subsets: ['latin'] });
 
 export const metadata = {
   metadataBase: new URL(OG_URL),
@@ -42,5 +52,27 @@ export const metadata = {
 };
 
 export default function RootLayout({ children }: { children: React.ReactNode }) {
-  return <Html locale="en">{children}</Html>;
+  return (
+    <html suppressHydrationWarning lang="en">
+      <head>
+        <meta
+          content="Level up your typescript skills with interactive exercises"
+          name="description"
+        />
+        <link href="/favicon.svg" rel="icon" type="image/svg+xml" />
+        <link rel="apple-touch-icon" sizes="180x180" href="/apple-touch-icon.png" />
+        <link rel="icon" type="image/png" sizes="32x32" href="/favicon-32x32.png" />
+        <link rel="icon" type="image/png" sizes="16x16" href="/favicon-16x16.png" />
+        <link rel="manifest" href="/site.webmanifest" />
+      </head>
+      <body className={`${inter.className} flex flex-col`}>
+        <Providers>
+          <Navigation />
+          <main className="flex-1">{children}</main>
+          <Toaster />
+        </Providers>
+        <Analytics />
+      </body>
+    </html>
+  );
 }
